[WebConsoleV2] Fixes the scheduler starting issue.
