+++ web_console_v2/api/server.py
@@ -13,18 +13,7 @@
 # limitations under the License.
 
 # coding: utf-8
-
+from config import Config
 from fedlearner_webconsole.app import create_app
-from fedlearner_webconsole.db import db
-from fedlearner_webconsole.auth.models import User
 
-app = create_app('config.Config')
-
-@app.cli.command('create-initial-data')
-def create_initial_data():
-    # Creates ada if it not exists
-    if User.query.filter_by(username='ada').first() is None:
-        user = User(username='ada')
-        user.set_password('ada')
-        db.session.add(user)
-        db.session.commit()
+app = create_app(Config())
